Added EmbedHandler, moved Timers and CommandHandler
