+++ CommunityBot/Handlers/CommandHandler.cs
@@ -1,10 +1,10 @@
-﻿using Discord.Commands;
-using Discord.WebSocket;
-using System;
+﻿using System;
 using System.Reflection;
 using System.Threading.Tasks;
+using Discord.Commands;
+using Discord.WebSocket;
 
-namespace CommunityBot
+namespace CommunityBot.Handlers
 {
     internal class CommandHandler
     {
