+++ CommunityBot/Helpers/Enums.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CommunityBot.Helpers
+{
+    class Enums
+    {
+    }
+
+    public enum EmbedMessageType
+    {
+        Success = 0,
+        Info = 10,
+        Error = 20,
+        Exception = 30
+    }
+}
