+++ CommunityBot/Helpers/Timers.cs
@@ -1,14 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
-using Discord;
-using Discord.Webhook;
 using System.Timers;
 using Discord.WebSocket;
 
-namespace CommunityBot
+namespace CommunityBot.Helpers
 {
     internal static class Timers
     {
